remove login.jsp from voter

+++ Electio/src/java/Action/ElectionResult.java
@@ -53,7 +53,7 @@ public class ElectionResult implements Controller.Action {
                     title = "Login";
                 } else {
                     ArrayList<Candidate> candidates = null;
-                    long election_type = (int) objE.getElectionType(election_id).getType_id();
+                    long election_type = (long) objE.getElectionType(election_id).getType_id();
                     req.setAttribute("election_type", election_type + "");
                     if (election_type == 1) {
                         candidates = objC.getCandidatesForPreferentialVoting(election_id);
